+++ modules/util/config/TrainConfig.py
@@ -71,7 +71,7 @@ class TrainOptimizerConfig(BaseConfig):
     nesterov: bool
     no_prox: bool
     optim_bits: int
-    percentile_clipping: float
+    percentile_clipping: int
     r: float
     relative_step: bool
     safeguard_warmup: bool
@@ -142,7 +142,7 @@ class TrainOptimizerConfig(BaseConfig):
         data.append(("nesterov", False, bool, False))
         data.append(("no_prox", False, bool, False))
         data.append(("optim_bits", None, int, True))
-        data.append(("percentile_clipping", None, float, True))
+        data.append(("percentile_clipping", None, int, True))
         data.append(("r", None, float, True))
         data.append(("relative_step", False, bool, False))
         data.append(("safeguard_warmup", False, bool, False))
